+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   move so that you can get reasons for each fence placement being impossible.
 * Fixed the default `sameMove` implementation.
 * The default `sameMove` was still not working for ambiguous ordo moves in Ordo. Implemented sortedReplacer.
+* Chase: Fix (and improve) `sameMove`.
 
 ### Changed